[PR] 수주

+++ Backend/erpc/src/main/java/com/cineverse/erpc/contract/aggregate/ContractProduct.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.contract.aggregate;
 
-import com.cineverse.erpc.product.aggregate.entity.Product;
+import com.cineverse.erpc.product.aggregate.Product;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.Data;
@@ -13,7 +13,7 @@ public class ContractProduct {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "contract_product_id")
-    private int contractProductId;
+    private long contractProductId;
 
     @Column(name = "contract_product_count")
     private int contractProductCount;
